refactored jsonapi files to api
